update donation
added card to kofi and changed title above patreon to h1. I would like to change the button to a normal one if we can't get the flowbite ones to work. I checked the app.css and I already formatted the normal buttons so it should be good to go